Write Voronoi cells to database instead of points
The grid that I wrote to the database was comprised of (buffered)
points. Now, the Voronoi cell containing a grid point is written to the
database instead, so you if one wants to know the weather for a certain
point, one doesn't have to do a nearest neighbour check on the whole
grid, but instead can check which cell contains the point of interest.
While at it, I also implemented that if a grid entry with a certain ID
already exists in the database, it's geometry will be updated, i.e.
OVERWRITTEN. Previously, grid entries already present in the database
would remain untouched. The reason for changing behaviour like this is
the fact, that anybody who used previous versions of the script,
probably has grid entries with incorrect geometries, which need to be
updated. Since this might destroy data, I added a notice to the
README.md file.